Move TableMetadataBuilder to TableMetadata and rename
